Warning: This builds and runs but eventually crashes
This is a first attempt at getting Bluetooth working. My idea for this branch is to write it
as a self-contained "Bluetooth test tool" which can later be turned into a self-contained
"controller" class in our main project. Currently I am stuck on the app crashing when I spawn
a worker thread. I'm not familiar with Android or Java threading specifically, so maybe the
way I'm currently trying to do threading is not proper.